requirements.txt
twisted==24.7.0
klein==23.5.0
txpostgres==1.7.0
psycopg2-binary==2.9.9
PyYAML==6.0.2
Jinja2==3.1.4
pytest==8.3.2
